Split multi-describe spec files into one file per class
