+++ spec/ammitto/sanction_entry_spec.rb
@@ -55,55 +55,3 @@ RSpec.describe Ammitto::SanctionEntry do
     end
   end
 end
-
-RSpec.describe Ammitto::Authority do
-  describe '.find' do
-    it 'returns authority for known code' do
-      auth = described_class.find('eu')
-      expect(auth).not_to be_nil
-      expect(auth.id).to eq('eu')
-      expect(auth.name).to eq('European Union')
-    end
-
-    it 'returns nil for unknown code' do
-      expect(described_class.find('xyz')).to be_nil
-    end
-  end
-
-  describe '.all' do
-    it 'returns all registered authorities' do
-      authorities = described_class.all
-      expect(authorities).to be_an(Array)
-      expect(authorities.size).to eq(16)
-    end
-
-    it 'covers every source the gem advertises' do
-      require 'ammitto/config/defaults'
-      missing = Ammitto::Config::Defaults::ALL_SOURCES.reject do |code|
-        described_class.find(code.to_s)
-      end
-      expect(missing).to be_empty
-    end
-  end
-end
-
-RSpec.describe Ammitto::TemporalPeriod do
-  it 'is active when no expiry date' do
-    period = described_class.new(listed_date: '2020-01-01', is_indefinite: true)
-    expect(period.active?).to be true
-  end
-
-  it 'is not active when expiry date passed' do
-    period = described_class.new(
-      listed_date: '2020-01-01',
-      expiry_date: '2020-12-31'
-    )
-    expect(period.active?).to be false
-  end
-
-  it 'calculates days until expiry' do
-    future_date = Date.today + 30
-    period = described_class.new(listed_date: '2020-01-01', expiry_date: future_date)
-    expect(period.days_until_expiry).to be_within(1).of(30)
-  end
-end
