+++ spec/ammitto/transformers/base_transformer_spec.rb
@@ -1,0 +1,48 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+RSpec.describe Ammitto::Transformers::BaseTransformer do
+  let(:transformer) { described_class.new(:test) }
+
+  describe '#source_code' do
+    it 'returns the source code' do
+      expect(transformer.source_code).to eq(:test)
+    end
+  end
+
+  describe '#transform' do
+    it 'raises NotImplementedError' do
+      expect { transformer.transform({}) }.to raise_error(NotImplementedError)
+    end
+  end
+
+  describe '#generate_entity_id' do
+    it 'generates a LIST-AGNOSTIC entity URI (sanitized to lowercase)' do
+      id = transformer.send(:generate_entity_id, 'TEST123')
+      # Entity IRIs are list-agnostic (no list_type) and sanitized to lowercase
+      expect(id).to eq('https://www.ammitto.org/entity/test/test123')
+    end
+  end
+
+  describe '#generate_entry_id' do
+    it 'generates a LIST-SPECIFIC entry URI' do
+      id = transformer.send(:generate_entry_id, 'TEST123')
+      # Entry IRIs are list-specific and include list_type
+      # For unknown source like :test, default_list_type returns nil which becomes 'unknown'
+      expect(id).to eq('https://www.ammitto.org/entry/test/unknown/test123')
+    end
+
+    it 'allows list_type override' do
+      id = transformer.send(:generate_entry_id, 'TEST123', entry_list_type: 'consolidated-list')
+      expect(id).to eq('https://www.ammitto.org/entry/test/consolidated-list/test123')
+    end
+  end
+
+  describe '#authority' do
+    it 'returns the authority for the source' do
+      auth = transformer.send(:authority)
+      expect(auth).to be_nil # :test is not a registered authority
+    end
+  end
+end
