+++ spec/ammitto/transformers/registry_spec.rb
@@ -79,228 +79,3 @@ RSpec.describe Ammitto::Transformers::Registry do
     end
   end
 end
-
-RSpec.describe Ammitto::Transformers::BaseTransformer do
-  let(:transformer) { described_class.new(:test) }
-
-  describe '#source_code' do
-    it 'returns the source code' do
-      expect(transformer.source_code).to eq(:test)
-    end
-  end
-
-  describe '#transform' do
-    it 'raises NotImplementedError' do
-      expect { transformer.transform({}) }.to raise_error(NotImplementedError)
-    end
-  end
-
-  describe '#generate_entity_id' do
-    it 'generates a LIST-AGNOSTIC entity URI (sanitized to lowercase)' do
-      id = transformer.send(:generate_entity_id, 'TEST123')
-      # Entity IRIs are list-agnostic (no list_type) and sanitized to lowercase
-      expect(id).to eq('https://www.ammitto.org/entity/test/test123')
-    end
-  end
-
-  describe '#generate_entry_id' do
-    it 'generates a LIST-SPECIFIC entry URI' do
-      id = transformer.send(:generate_entry_id, 'TEST123')
-      # Entry IRIs are list-specific and include list_type
-      # For unknown source like :test, default_list_type returns nil which becomes 'unknown'
-      expect(id).to eq('https://www.ammitto.org/entry/test/unknown/test123')
-    end
-
-    it 'allows list_type override' do
-      id = transformer.send(:generate_entry_id, 'TEST123', entry_list_type: 'consolidated-list')
-      expect(id).to eq('https://www.ammitto.org/entry/test/consolidated-list/test123')
-    end
-  end
-
-  describe '#authority' do
-    it 'returns the authority for the source' do
-      auth = transformer.send(:authority)
-      expect(auth).to be_nil # :test is not a registered authority
-    end
-  end
-end
-
-RSpec.describe Ammitto::Sources::Uk::Transformer do
-  let(:transformer) { described_class.new }
-
-  describe '#source_code' do
-    it 'returns :uk' do
-      expect(transformer.source_code).to eq(:uk)
-    end
-  end
-
-  describe '#authority' do
-    it 'returns UK authority' do
-      auth = transformer.send(:authority)
-      expect(auth.id).to eq('uk')
-      expect(auth.name).to eq('United Kingdom (OFSI)')
-    end
-  end
-end
-
-RSpec.describe Ammitto::Sources::Eu::Transformer do
-  let(:transformer) { described_class.new }
-
-  describe '#source_code' do
-    it 'returns :eu' do
-      expect(transformer.source_code).to eq(:eu)
-    end
-  end
-
-  describe '#authority' do
-    it 'returns EU authority' do
-      auth = transformer.send(:authority)
-      expect(auth.id).to eq('eu')
-      expect(auth.name).to eq('European Union')
-    end
-  end
-end
-
-RSpec.describe Ammitto::Sources::Un::Transformer do
-  let(:transformer) { described_class.new }
-
-  describe '#source_code' do
-    it 'returns :un' do
-      expect(transformer.source_code).to eq(:un)
-    end
-  end
-
-  describe '#authority' do
-    it 'returns UN authority' do
-      auth = transformer.send(:authority)
-      expect(auth.id).to eq('un')
-      expect(auth.name).to eq('United Nations')
-    end
-  end
-end
-
-RSpec.describe Ammitto::Sources::Us::Transformer do
-  let(:transformer) { described_class.new }
-
-  describe '#source_code' do
-    it 'returns :us' do
-      expect(transformer.source_code).to eq(:us)
-    end
-  end
-
-  describe '#authority' do
-    it 'returns US authority' do
-      auth = transformer.send(:authority)
-      expect(auth.id).to eq('us')
-      expect(auth.name).to eq('United States (OFAC)')
-    end
-  end
-end
-
-RSpec.describe Ammitto::Sources::Wb::Transformer do
-  let(:transformer) { described_class.new }
-
-  describe '#source_code' do
-    it 'returns :wb' do
-      expect(transformer.source_code).to eq(:wb)
-    end
-  end
-
-  describe '#authority' do
-    it 'returns WB authority' do
-      auth = transformer.send(:authority)
-      expect(auth.id).to eq('wb')
-      expect(auth.name).to eq('World Bank')
-    end
-  end
-end
-
-RSpec.describe Ammitto::Sources::Au::Transformer do
-  let(:transformer) { described_class.new }
-
-  describe '#source_code' do
-    it 'returns :au' do
-      expect(transformer.source_code).to eq(:au)
-    end
-  end
-
-  describe '#authority' do
-    it 'returns AU authority' do
-      auth = transformer.send(:authority)
-      expect(auth.id).to eq('au')
-      expect(auth.name).to eq('Australia (DFAT)')
-    end
-  end
-end
-
-RSpec.describe Ammitto::Sources::Ca::Transformer do
-  let(:transformer) { described_class.new }
-
-  describe '#source_code' do
-    it 'returns :ca' do
-      expect(transformer.source_code).to eq(:ca)
-    end
-  end
-
-  describe '#authority' do
-    it 'returns CA authority' do
-      auth = transformer.send(:authority)
-      expect(auth.id).to eq('ca')
-      expect(auth.name).to eq('Canada (SEFO)')
-    end
-  end
-end
-
-RSpec.describe Ammitto::Sources::Ch::Transformer do
-  let(:transformer) { described_class.new }
-
-  describe '#source_code' do
-    it 'returns :ch' do
-      expect(transformer.source_code).to eq(:ch)
-    end
-  end
-
-  describe '#authority' do
-    it 'returns CH authority' do
-      auth = transformer.send(:authority)
-      expect(auth.id).to eq('ch')
-      expect(auth.name).to eq('Switzerland (SECO)')
-    end
-  end
-end
-
-RSpec.describe Ammitto::Sources::Cn::Transformer do
-  let(:transformer) { described_class.new }
-
-  describe '#source_code' do
-    it 'returns :cn' do
-      expect(transformer.source_code).to eq(:cn)
-    end
-  end
-
-  describe '#authority' do
-    it 'returns CN authority' do
-      auth = transformer.send(:authority)
-      expect(auth.id).to eq('cn')
-      expect(auth.name).to eq('China (MOFCOM/MFA)')
-    end
-  end
-end
-
-RSpec.describe Ammitto::Sources::Ru::Transformer do
-  let(:transformer) { described_class.new }
-
-  describe '#source_code' do
-    it 'returns :ru' do
-      expect(transformer.source_code).to eq(:ru)
-    end
-  end
-
-  describe '#authority' do
-    it 'returns RU authority' do
-      auth = transformer.send(:authority)
-      expect(auth.id).to eq('ru')
-      expect(auth.name).to eq('Russia (MID/CBR)')
-    end
-  end
-end
